Refactor: import package명 수정 및 서비스 로직 수정
