+++ src/main/java/com/team4/goorm/community/Member/domain/SocialType.java
@@ -1,4 +1,4 @@
-package com.team4.goorm.community.member.domain;
+package com.team4.goorm.community.Member.domain;
 
 public enum SocialType {
 	KAKAO,
